Issue-3499: Move progress_stream module to compose/common package
Partially implements #3499
