+++ compose/common/progress_stream.py
